+++ lib/dbf/format_profile.ex
@@ -9,6 +9,7 @@ defmodule DBF.FormatProfile do
     :header_layout,
     :field_descriptor_layout,
     :memo_family,
+    :memo_requirement,
     :record_layout,
     :field_kinds
   ]
@@ -18,16 +19,27 @@ defmodule DBF.FormatProfile do
     :header_layout,
     :field_descriptor_layout,
     :memo_family,
+    :memo_requirement,
     :record_layout,
     :field_kinds
   ]
 
-  @type header_layout :: :foxbase_8 | :dbase_legacy_32
-  @type field_descriptor_layout :: :foxbase_16 | :dbase_legacy_32
+  @type header_layout :: :foxbase_8 | :dbase_legacy_32 | :visual_foxpro_32
+  @type field_descriptor_layout :: DBF.FieldDescriptorLayout.t()
   @type memo_family :: :none | :dbt_iii | :dbt_iv | :fpt
+  @type memo_requirement :: :none | :required | :table_flag
   @type record_layout :: :dbase_legacy
   @type field_kind ::
-          :character | :numeric | :numeric_unscaled | :float | :logical | :date | :text_memo
+          :character
+          | :numeric
+          | :numeric_unscaled
+          | :float
+          | :logical
+          | :date
+          | :text_memo
+          | :integer
+          | :datetime
+          | :text_memo_binary_pointer
 
   @type t :: %__MODULE__{
           version: byte(),
@@ -35,6 +47,7 @@ defmodule DBF.FormatProfile do
           header_layout: header_layout(),
           field_descriptor_layout: field_descriptor_layout(),
           memo_family: memo_family(),
+          memo_requirement: memo_requirement(),
           record_layout: record_layout(),
           field_kinds: %{binary() => field_kind()}
         }
@@ -48,6 +61,11 @@ defmodule DBF.FormatProfile do
   }
 
   @memo_field_kinds Map.put(@legacy_field_kinds, "M", :text_memo)
+  @visual_foxpro_field_kinds Map.merge(@legacy_field_kinds, %{
+                               "I" => :integer,
+                               "M" => :text_memo_binary_pointer,
+                               "T" => :datetime
+                             })
 
   @profiles %{
     0x02 => %{
@@ -56,6 +74,7 @@ defmodule DBF.FormatProfile do
       header_layout: :foxbase_8,
       field_descriptor_layout: :foxbase_16,
       memo_family: :none,
+      memo_requirement: :none,
       record_layout: :dbase_legacy,
       field_kinds: %{"C" => :character, "N" => :numeric_unscaled}
     },
@@ -65,8 +84,19 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :none,
+      memo_requirement: :none,
       record_layout: :dbase_legacy,
       field_kinds: @legacy_field_kinds
+    },
+    0x30 => %{
+      version: 0x30,
+      label: "Visual FoxPro",
+      header_layout: :visual_foxpro_32,
+      field_descriptor_layout: :visual_foxpro_32,
+      memo_family: :fpt,
+      memo_requirement: :table_flag,
+      record_layout: :dbase_legacy,
+      field_kinds: @visual_foxpro_field_kinds
     },
     0x83 => %{
       version: 0x83,
@@ -74,6 +104,7 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :dbt_iii,
+      memo_requirement: :required,
       record_layout: :dbase_legacy,
       field_kinds: @memo_field_kinds
     },
@@ -83,6 +114,7 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :dbt_iv,
+      memo_requirement: :required,
       record_layout: :dbase_legacy,
       field_kinds: @memo_field_kinds
     },
@@ -92,6 +124,7 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :fpt,
+      memo_requirement: :required,
       record_layout: :dbase_legacy,
       field_kinds: @memo_field_kinds
     }